+++ src/Essentials/Weapon.java
@@ -2,7 +2,7 @@ package Essentials;
 
 public class Weapon extends MainStats {
 
-    public Weapon(String tag, double hp, double atk, double def, double crit) {
+    public Weapon(String tag, int hp, int atk, int def, int crit) {
         super(tag, hp, atk, def, crit);
     }
 }
